midaircommit

+++ src/main/java/no/nav/bidrag/beregn/samvaersfradrag/beregning/SamvaersfradragBeregningImpl.java
@@ -1,5 +1,6 @@
 package no.nav.bidrag.beregn.samvaersfradrag.beregning;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import no.nav.bidrag.beregn.felles.SjablonUtil;
@@ -20,7 +21,7 @@ public class SamvaersfradragBeregningImpl implements SamvaersfradragBeregning {
 
     List<SjablonNokkel> sjablonNokkelListe = new ArrayList<>();
 
-    double belopFradrag = 0.0d;
+    BigDecimal belopFradrag = BigDecimal.valueOf(0);
 
     sjablonNokkelListe.add(new SjablonNokkel(SjablonNokkelNavn.SAMVAERSKLASSE.getNavn(),
           grunnlagBeregningPeriodisert.getSamvaersklasse()));
